Implement batch endpoint for todo server app.
* Create a generic "mutator" pattern as an analogue to existing "handlers"
* Change existing CreateTodo handler to mutator instead
* Reimplement CreateTodo handler in terms of mutator (useful to have
  non-Replicache handler around for demos)

Still left todo (heh):

* Implement the other mutators
* Not super thrilled with all the new error types, is there some way
  to do same thing with less code/boilerplate?

Toward rocicorp/replicache#30

+++ serve/model/replicache/replicache.go
@@ -22,7 +22,7 @@ func GetMutationID(d *db.DB, clientID string) (int64, error) {
 		return int64(0), fmt.Errorf("unexpected number of MutationID records")
 	}
 
-	return int64(*output.Records[0][0].LongValue), nil
+	return *output.Records[0][0].LongValue, nil
 }
 
 // SetMutationID updates the database with the last processed clientID for a client.
